requirements.txt
# Requirements for running the program
numpy
requests
Pillow>=4.2.0
#keras-retinanet==0.4.1
git+https://github.com/fizyr/keras-retinanet.git@08af308d01a8f22dc286d62bc26c8496e1ff6539
readlif>=0.2.0
tensorflow==1.15.2
sk-video
opencv-python
PySimpleGUI
pandas
keras==2.2.5
discover